requirements.txt
flask==2.3.2
flask-cors==4.0.0
tensorflow-cpu==2.13.0
pandas==2.2.2
joblib==1.3.2
scikit-learn==1.3.2
gunicorn==21.2.0
